updating blueprints to files
